+++ ostap/io/bz2shelve.py
@@ -254,7 +254,7 @@ class Bz2Shelf(CompressShelf):
                 return tuple ( items )
             
         import tempfile , io   
-        fd , fileout = tempfile.mkstemp ( prefix = 'ostap-tmp-' , suffix = '-db' )
+        fd , fileout = tempfile.mkstemp ( prefix = 'ostap-tmp-' , suffix = '-bz2db' )
         with bz2.open ( filein  , 'rb' ) as fin : 
             with io.open ( fileout , 'wb' ) as fout : 
                 shutil.copyfileobj ( fin , fout )                
@@ -302,12 +302,8 @@ class Bz2Shelf(CompressShelf):
                             keyencoding = self.keyencoding   )
         
         ## copy the content
-        if keys :
-            for key in self.keys() :
-                if key in keys     : new_db [ key ] = self [ key ]
-        else : 
-            for key in self.keys() : new_db [ key ] = self [ key ]
-            
+        copy = keys if keys else self.keys()
+        for key in copy : new_db [ key ] = self [ key ]
         new_db.sync ()  
         return new_db