+++ ReleaseNotes/release_notes.md
@@ -1,5 +1,7 @@
 ## New features:
 
+  1. add new module `ostap.io.zstshelve` with shelve-like databse using `zstandard` compression 
+
 ## Backward incompatible: 
 
 ## Bug fixes: